requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.6
djangorestframework==3.14.0
exceptiongroup==1.1.0
Faker==12.0.1
iniconfig==2.0.0
mixer==7.2.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
